Reorganize code so allmydata.web.check_results can import without Nevow being installed.
